[BE] FIX: 길드 기본메시지 api 수정

+++ backend/backend/src/main/java/com/crimecat/backend/guild/service/ChannelRecordService.java
@@ -5,8 +5,12 @@ import com.crimecat.backend.guild.dto.ChannelRecordDto;
 import com.crimecat.backend.guild.dto.ChannelRecordListResponseDto;
 import com.crimecat.backend.guild.dto.ChannelRecordRequestDto;
 import com.crimecat.backend.guild.repository.ChannelRecordRepository;
+import jakarta.transaction.Transactional;
+import jakarta.validation.ConstraintViolationException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Comparator;
 import java.util.Optional;
@@ -15,8 +19,12 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ChannelRecordService {
     private final ChannelRecordRepository channelRecordRepository;
+    private final GuildQueryService guildQueryService;
 
     public ChannelRecordListResponseDto getRecords(String guildSnowflake) {
+        if (!guildQueryService.existsBySnowflake(guildSnowflake)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "guild not exists");
+        }
         return new ChannelRecordListResponseDto(
                 channelRecordRepository.findByGuildSnowflake(guildSnowflake).stream()
                         .sorted(Comparator.comparingInt(Record::getIndex))
@@ -26,6 +34,10 @@ public class ChannelRecordService {
     }
 
     public void addRecord(String guildSnowflake, ChannelRecordRequestDto channelRecordRequestDto) {
+        if (!guildQueryService.existsBySnowflake(guildSnowflake)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "guild not exists");
+        }
+        // TODO: getLastIndex()
         Optional<Record> record = channelRecordRepository.findByGuildSnowflakeAndChannelSnowflake(
                 guildSnowflake, channelRecordRequestDto.getChannelSnowflake()
         ).stream()
@@ -36,14 +48,31 @@ public class ChannelRecordService {
         if (record.isPresent()) {
             index = record.get().getIndex() + 1;
         }
-        channelRecordRepository.save(new Record(channelRecordRequestDto, index, guildSnowflake));
+        Record newRecord = new Record(channelRecordRequestDto, index, guildSnowflake);
+        // TODO: ConstraintViolationException 처리 (validation 처리)
+        try {
+            channelRecordRepository.save(newRecord);
+        } catch (Exception e) {
+            Throwable t = e.getCause();
+            while (t != null && !(t instanceof ConstraintViolationException)) {
+                t = t.getCause();
+            }
+            if (t != null) {
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "body require element need");
+            }
+            throw e;
+        }
     }
 
+    @Transactional
     public void deleteRecord(String guildSnowflake, String channelSnowflake) {
+        if (!guildQueryService.existsBySnowflake(guildSnowflake)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "guild not exists");
+        }
         long deletedNum = channelRecordRepository.deleteByGuildSnowflakeAndChannelSnowflake(guildSnowflake,
                 channelSnowflake);
         if (deletedNum == 0) {
-            throw new RuntimeException();
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "record not exists");
         }
     }
 }
